feat: Add ML anomaly detection with explanation feature
Backend ML Engine:
- Add Isolation Forest-based anomaly detection system
- Implement feature importance calculation using perturbation method
- Add training statistics tracking (mean, std, IQR for each feature)
- Generate Korean explanation for detected anomalies
- Create ML analysis API endpoints (train, analyze, statistics, models)

Backend Changes:
- backend/app/ml/: Complete ML engine implementation
  - anomaly_detector.py: Isolation Forest with feature importance
  - predictor.py: Prediction with auto-generated explanations
  - preprocessor.py: Data preprocessing with statistics tracking
  - trainer.py: Model training logic
  - utils.py: Utility functions
- backend/app/services/ml_service.py: ML orchestration service
- backend/app/routers/ml_analysis.py: ML API endpoints
- backend/app/schemas/ml_analysis.py: ML request/response schemas with explanation field
- backend/app/main.py: Register ML router and update CORS
- backend/requirements.txt: Add scikit-learn, pandas, numpy, joblib

Frontend ML Dashboard:
- Complete ML visualization with 5 pages
- Display anomaly explanations in "위험 원인" column
- Parse multiple reasons separated by " | " into bullet list
- Highlight risk factors in red with proper styling

Frontend Changes:
- frontend/src/app/ml/: ML dashboard pages
  - page.tsx: Main ML dashboard
  - train/page.tsx: Model training interface
  - analyze/page.tsx: Log analysis with explanation display
  - statistics/page.tsx: Statistics visualization
  - models/page.tsx: Model management
- frontend/src/components/ml/: ML components
  - AnalysisResults.tsx: Results table with explanation column
  - TrainingForm.tsx: Model training form
  - ModelCard.tsx: Model information card
  - AnomalyBadge.tsx: Anomaly indicator badge
- frontend/src/components/charts/: Chart components
  - AnomalyScoreGauge.tsx: Score gauge
  - TimeSeriesChart.tsx: Time series visualization
  - BarChart.tsx: Bar chart component
  - StatisticsCard.tsx: Statistics card
- frontend/src/lib/mlApi.ts: ML API client functions
- frontend/src/types/ml.ts: ML TypeScript types with explanation field
- frontend/src/components/Navigation.tsx: Add ML navigation link
- frontend/package.json: Add recharts for visualization

Other Changes:
- logcollector/db.c: Fix compilation errors (add stdlib.h, fix time_t format)
- .claude/docs/progress.md: Update progress log with ML explanation feature

Features:
- Quantitative anomaly analysis with IQR-based thresholds
- Automatic explanation generation in Korean
- Feature importance ranking to identify key anomaly factors
- User-friendly visualization with color-coded risk indicators
- Multi-reason parsing and display for comprehensive insights

+++ logcollector/db.c
@@ -1,6 +1,7 @@
 #include "db.h"
 #include "config.h"
 #include <stdio.h>
+#include <stdlib.h>
 #include <string.h>
 #include <sys/stat.h>
 #include <time.h>
@@ -119,7 +120,7 @@ int db_check_and_roll(LogDB *db, int max_size_mb) {
     // Generate new DB filename with timestamp
     char new_path[512];
     time_t now = time(NULL);
-    snprintf(new_path, sizeof(new_path), "%s%ld.db", DB_ROLLING_PREFIX, now);
+    snprintf(new_path, sizeof(new_path), "%s%lld.db", DB_ROLLING_PREFIX, (long long)now);
 
     // Rename current DB
     rename(db->db_path, new_path);
